+++ app/workers/telegram_group_auto_close_worker.rb
@@ -16,14 +16,16 @@ class TelegramGroupAutoCloseWorker
   end
 
   def need_to_notify_issues
+    issues = Issue.joins(:telegram_group)
+                  .where('redmine_chat_telegram_telegram_groups.last_notification_at <= ?', 24.hours.ago.change(min: 59, sec: 59))
+
     if close_issue_status_ids.present?
-      Issue.joins(:telegram_group)
-        .where(status_id: close_issue_status_ids)
-        .where('redmine_chat_telegram_telegram_groups.last_notification_at <= ?', 12.hours.ago)
+      issues = issues.where(status_id: close_issue_status_ids)
     else
-      Issue.open(false).joins(:telegram_group)
-        .where('redmine_chat_telegram_telegram_groups.last_notification_at <= ?', 12.hours.ago)
+      issues = issues.open(false)
     end
+
+    issues
   end
 
   def close_old_chats