+++ app/workers/telegram_handler_worker.rb
@@ -3,10 +3,18 @@ class TelegramHandlerWorker
   sidekiq_options queue: :telegram
 
   def perform(params)
-    message = Telegram::Bot::Types::Update.new(params).message
+    update = Telegram::Bot::Types::Update.new(params)
+
+    types = %w(inline_query
+                   chosen_inline_result
+                   callback_query
+                   edited_message
+                   message
+                   channel_post
+                   edited_channel_post)
+    message = types.inject(nil) { |acc, elem| acc || update.send(elem) }
 
     if message.present?
-      logger.debug params.to_json
       RedmineChatTelegram.handle_message(message)
     else
       logger.fatal "Can't find message: #{params.to_json}"