+++ init.rb
@@ -4,6 +4,7 @@ FileUtils.mkdir_p(Rails.root.join('log/chat_telegram')) unless Dir.exist?(Rails.
 
 require 'pluralization'
 require 'redmine_chat_telegram'
+require 'telegram/bot'
 
 ActionDispatch::Callbacks.to_prepare do
   paths = '/lib/redmine_chat_telegram/{patches/*_patch,hooks/*_hook}.rb'
@@ -26,7 +27,7 @@ Redmine::Plugin.register :redmine_chat_telegram do
   name 'Redmine Chat Telegram plugin'
   url 'https://github.com/centosadmin/redmine_chat_telegram'
   description 'This is a plugin for Redmine which adds Telegram Group Chat to Redmine Issue'
-  version '2.0.0'
+  version '2.0.1-dev'
   author 'Southbridge'
   author_url 'https://github.com/centosadmin'