+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### [2.0.1] (In progress)
+
+* Now edited messages correctly replacing in chat archive
+* Add message about access denied when not connected user try to create new issue
+* Fix pluralization for en and ru close notification messages
+
 #### [2.0.0]
 
 * Redmine user which have proper permissions and joined to group will became administrator of the group